style(components): move types and helpers
remove analytics stories

J=SLAP=2399
TEST=manual

ran storybook
