+++ packages/pages/src/components/link/link.tsx
@@ -1,73 +1,8 @@
-import { ConversionDetails } from "@yext/analytics";
 import React, { useState } from "react";
 import classNames from "classnames";
 import { useAnalytics } from "../analytics/index.js";
-import { getHref, isEmail, reverse } from "./methods.js";
-import type { CTA } from "./types.js";
-
-/**
- * Configuration options available for any usages of the Link component.
- *
- * @public
- */
-interface LinkConfig
-  extends React.DetailedHTMLProps<
-    React.AnchorHTMLAttributes<HTMLAnchorElement>,
-    HTMLAnchorElement
-  > {
-  obfuscate?: boolean;
-  eventName?: string;
-  conversionDetails?: ConversionDetails | undefined;
-}
-
-/**
- * The shape of the data passed to {@link Link} when directly passing an HREF to the Link component.
- */
-export interface HREFLinkProps extends LinkConfig {
-  href: string;
-  cta?: never;
-}
-
-/**
- * The shape of the data passed to {@link Link} when using a CTA field, and not overriding children.
- */
-export interface CTAWithChildrenLinkProps extends LinkConfig {
-  href?: never;
-  cta: CTA;
-  children?: never;
-}
-
-/**
- * The shape of the data passed to {@link Link} when using a CTA field, and overriding children.
- *
- * @public
- */
-export interface CTAWithoutChildrenLinkProps extends LinkConfig {
-  href?: never;
-  cta: Omit<CTA, "label">;
-  children: React.ReactNode;
-}
-
-/**
- * The shape of the data passed to {@link Link} when using a CTA field.
- *
- * @public
- */
-type CTALinkProps = CTAWithChildrenLinkProps | CTAWithoutChildrenLinkProps;
-
-/**
- * The shape of the data passed to {@link Link}.
- *
- * @public
- */
-export type LinkProps = CTALinkProps | HREFLinkProps;
-
-/**
- * Type predicate for distinguishing between data cases.
- */
-function isHREFProps(props: LinkProps): props is HREFLinkProps {
-  return "href" in props;
-}
+import { getHref, isEmail, isHREFProps, reverse } from "./methods.js";
+import type { CTA, LinkProps } from "./types.js";
 
 /**
  * Renders an anchor tag using either a directly provided HREF or from a field in the Yext Knowledge Graph.
